chore: bring back js files via using type
