+++ packages/gatsby-theme-carbon/package.json
@@ -3,6 +3,7 @@
   "version": "4.0.0-next.4",
   "main": "index.js",
   "author": "vpicone <vp@vincepic.one> (@vpicone)",
+  "type": "module",
   "repository": {
     "url": "git@github.com:carbon-design-system/gatsby-theme-carbon.git",
     "type": "git"
